Fixing multiplication result converted to larger type reports in incremental tests
Fix https://github.com/kokkos/kokkos/security/code-scanning/17
https://github.com/kokkos/kokkos/security/code-scanning/16
https://github.com/kokkos/kokkos/security/code-scanning/15
https://github.com/kokkos/kokkos/security/code-scanning/14
https://github.com/kokkos/kokkos/security/code-scanning/13
https://github.com/kokkos/kokkos/security/code-scanning/12
https://github.com/kokkos/kokkos/security/code-scanning/11
https://github.com/kokkos/kokkos/security/code-scanning/10

+++ core/unit_test/incremental/Test12a_ThreadScratch.hpp
@@ -87,8 +87,8 @@ struct ThreadScratch {
     Kokkos::fence();
     auto v_H = Kokkos::create_mirror_view_and_copy(Kokkos::HostSpace(), v);
 
-    size_t check   = 0;
-    const size_t s = pN * sX * sY;
+    int check   = 0;
+    const int s = pN * sX * sY;
     for (int n = 0; n < pN; ++n)
       for (int m = 0; m < max_team_size; ++m) {
         check += v_H(n, m);
